+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- The client now gives a ephemeral timed out response when an unknown message is received.
+
+### Fixed
+- Handling of access errors in the component client.
 
 ## [1.0.1a1]
 ### Added